refactor: package structure reformed
