+++ src/main/java/shop/project/pathorderserver/_core/config/WebMvcConfig.java
@@ -12,6 +12,7 @@ public class WebMvcConfig implements WebMvcConfigurer {
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(new LoginInterceptor())
                 //.addPathPatterns("/api/**") // TODO: 로그인 및 JWT 검증시, 주석 제거
+                //.addPathPatterns("/stores/**") // TODO: 로그인 및 JWT 검증시, 주석 제거
                 .excludePathPatterns("/**");
     }
 }
