+++ src/main/java/shop/project/pathorderserver/order/OrderService.java
@@ -8,6 +8,7 @@ import shop.project.pathorderserver.store.Store;
 import shop.project.pathorderserver.store.StoreRepository;
 import shop.project.pathorderserver.user.User;
 import shop.project.pathorderserver.user.UserRepository;
+import shop.project.pathorderserver.user.UserResponse;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -21,7 +22,7 @@ public class OrderService {
     private final UserRepository userRepository;
     private final StoreRepository storeRepository;
 
-    @Transactional
+    @Transactional // 주문하기
     public OrderResponse.OrderDTO createOrder(OrderRequest.OrderDTO reqDTO) {
         User customer // 유저 번호로 유저 조회
                 = userRepository.findById(reqDTO.getCustomerId())
@@ -54,6 +55,36 @@ public class OrderService {
         orderRepository.save(order); // DB insert
 
         return new OrderResponse.OrderDTO(order, orderMenus, orderMenuOptions);
+    }
+
+    // 주문내역 목록보기 (손님)
+    public UserResponse.OrderListDTO getOrderList(int userId) {
+        List<Order> orders = orderRepository.findByUserId(userId)
+                .orElseThrow(() -> new Exception404("찾을 수 없는 주문입니다."));
+
+        return new UserResponse.OrderListDTO(orders);
+    }
+
+    @Transactional // 주문내역 상세보기 (손님)
+    public UserResponse.OrderDetailDTO getOrderDetail(int orderId) {
+        Order order // 단일 주문 내역 조회
+                = orderRepository.findById(orderId)
+                .orElseThrow(() -> new Exception404("찾을 수 없는 주문입니다."));
+        List<OrderMenu> orderMenus // 주문 내역의 메뉴 목록.
+                = orderMenuRepository.findAllByOrderId(orderId)
+                .orElseThrow(() -> new Exception404("찾을 수 없는 주문입니다."));
+        /* 양방향 매핑으로 변경.
+        List<Integer> orderMenuIdList; // 34, 35, 36
+        orderMenuIdList // 주문 메뉴별 Id 추출
+                = orderMenus.stream().map(orderMenu -> orderMenu.getId()).toList();
+        for (int orderMenuId : orderMenuIdList) {
+            List<OrderOption> orderOptions // 주문 메뉴별 옵션 목록 조회
+                    = orderMenuRepository.findAllOptionByMenuId(orderMenuId)
+                            .orElse(null); // 없으면 null
+            orderMenus.get(orderMenuId).setOrderOption(orderOptions);
+        }
+        */
+        return new UserResponse.OrderDetailDTO(order, orderMenus);
     }
     // TODO: 주문 수정 (매장 측: 주문 상태 변경)
 }
