+++ src/main/java/shop/project/pathorderserver/user/UserService.java
@@ -7,20 +7,19 @@ import shop.project.pathorderserver._core.errors.exception.Exception400;
 import shop.project.pathorderserver._core.errors.exception.Exception401;
 import shop.project.pathorderserver._core.errors.exception.Exception404;
 import shop.project.pathorderserver._core.utils.JwtUtil;
-import shop.project.pathorderserver.order.*;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.*;
+import java.util.Base64;
+import java.util.Optional;
+import java.util.UUID;
 
 @RequiredArgsConstructor
 @Service
 public class UserService {
     private final UserRepository userRepository;
-    private final OrderRepository orderRepository;
-    private final OrderMenuRepository orderMenuRepository;
 
     @Transactional // 회원 가입
     public UserResponse.JoinDTO createUser(UserRequest.JoinDTO reqDTO) {
@@ -96,35 +95,5 @@ public class UserService {
         user.setImgFilename(newFilename);
 
         return new UserResponse.ImgDTO(newFilePath.toString());
-    }
-
-    // 주문내역 목록보기
-    public UserResponse.OrderListDTO getOrderList(int userId) {
-        List<Order> orders = orderRepository.findByUserId(userId)
-                .orElseThrow(() -> new Exception404("찾을 수 없는 주문입니다."));
-
-        return new UserResponse.OrderListDTO(orders);
-    }
-
-    @Transactional // 주문내역 상세보기
-    public UserResponse.OrderDetailDTO getOrderDetail(int orderId) {
-        List<Integer> orderMenuIdList; // 34, 35, 36
-        Order order // 단일 주문 내역 조회
-                = orderRepository.findById(orderId)
-                .orElseThrow(() -> new Exception404("찾을 수 없는 주문입니다."));
-        List<OrderMenu> orderMenus // 주문 내역의 메뉴 목록.
-                = orderMenuRepository.findAllByOrderId(orderId)
-                .orElseThrow(() -> new Exception404("찾을 수 없는 주문입니다."));
-        /* 양방향 매핑으로 변경.
-        orderMenuIdList // 주문 메뉴별 Id 추출
-                = orderMenus.stream().map(orderMenu -> orderMenu.getId()).toList();
-        for (int orderMenuId : orderMenuIdList) {
-            List<OrderOption> orderOptions // 주문 메뉴별 옵션 목록 조회
-                    = orderMenuRepository.findAllOptionByMenuId(orderMenuId)
-                            .orElse(null); // 없으면 null
-            orderMenus.get(orderMenuId).setOrderOption(orderOptions);
-        }
-        */
-        return new UserResponse.OrderDetailDTO(order, orderMenus);
     }
 }
